requirements.txt
requests==2.31.0
openpyxl==3.1.2
python-dotenv==1.0.0
flask==3.0.0
